Combine cli package into one file
